# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface CreateCampaignBody {
  name: string;
  targetUrl: string;
  competitorUrls: string[];
  seedQueries: string[];
}

export type CampaignStatus =
  (typeof CampaignStatus)[keyof typeof CampaignStatus];

export const CampaignStatus = {
  idle: "idle",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface Campaign {
  id: number;
  name: string;
  targetUrl: string;
  competitorUrls: string[];
  seedQueries: string[];
  status: CampaignStatus;
  createdAt: string;
  completedAt?: string | null;
}

export interface LlmResponse {
  id: number;
  campaignId: number;
  llm: string;
  query: string;
  responseText: string;
  createdAt: string;
}

export interface BrandMentionStat {
  brand: string;
  url: string;
  mentions: number;
  totalMentions: number;
  shareOfVoice: number;
  avgRankPosition?: number | null;
  isTarget: boolean;
}

export type LlmVisibilityCompetitorBreakdownItem = {
  brand: string;
  mentions: number;
};

export interface LlmVisibility {
  llm: string;
  targetMentions: number;
  totalMentions: number;
  shareOfVoice: number;
  competitorBreakdown: LlmVisibilityCompetitorBreakdownItem[];
}

export interface AeoReport {
  campaignId: number;
  totalQueries: number;
  totalResponses: number;
  brandStats: BrandMentionStat[];
  llmVisibility: LlmVisibility[];
}