# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AEO Analytics Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: campaigns
    description: AEO analysis campaigns
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a new campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /campaigns/{id}/run:
    post:
      operationId: runCampaign
      tags: [campaigns]
      summary: Run the campaign analysis (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SSE stream of run progress
          content:
            text/event-stream:
              schema:
                type: string

  /campaigns/{id}/results:
    get:
      operationId: getCampaignResults
      tags: [campaigns]
      summary: Get stored LLM responses for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of LLM responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LlmResponse"

  /campaigns/{id}/report:
    get:
      operationId: getCampaignReport
      tags: [campaigns]
      summary: Get computed AEO metrics report for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AEO metrics report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AeoReport"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    CreateCampaignBody:
      type: object
      properties:
        name:
          type: string
        targetUrl:
          type: string
        competitorUrls:
          type: array
          items:
            type: string
        seedQueries:
          type: array
          items:
            type: string
      required:
        - name
        - targetUrl
        - competitorUrls
        - seedQueries

    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        targetUrl:
          type: string
        competitorUrls:
          type: array
          items:
            type: string
        seedQueries:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [idle, running, completed, failed]
        createdAt:
          type: string
        completedAt:
          type: string
          nullable: true
      required:
        - id
        - name
        - targetUrl
        - competitorUrls
        - seedQueries
        - status
        - createdAt

    LlmResponse:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        llm:
          type: string
        query:
          type: string
        responseText:
          type: string
        createdAt:
          type: string
      required:
        - id
        - campaignId
        - llm
        - query
        - responseText
        - createdAt

    BrandMentionStat:
      type: object
      properties:
        brand:
          type: string
        url:
          type: string
        mentions:
          type: integer
        totalMentions:
          type: integer
        shareOfVoice:
          type: number
        avgRankPosition:
          type: number
          nullable: true
        isTarget:
          type: boolean
      required:
        - brand
        - url
        - mentions
        - totalMentions
        - shareOfVoice
        - isTarget

    LlmVisibility:
      type: object
      properties:
        llm:
          type: string
        targetMentions:
          type: integer
        totalMentions:
          type: integer
        shareOfVoice:
          type: number
        competitorBreakdown:
          type: array
          items:
            type: object
            properties:
              brand:
                type: string
              mentions:
                type: integer
            required:
              - brand
              - mentions
      required:
        - llm
        - targetMentions
        - totalMentions
        - shareOfVoice
        - competitorBreakdown

    AeoReport:
      type: object
      properties:
        campaignId:
          type: integer
        totalQueries:
          type: integer
        totalResponses:
          type: integer
        brandStats:
          type: array
          items:
            $ref: "#/components/schemas/BrandMentionStat"
        llmVisibility:
          type: array
          items:
            $ref: "#/components/schemas/LlmVisibility"
      required:
        - campaignId
        - totalQueries
        - totalResponses
        - brandStats
        - llmVisibility