# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all campaigns
 */
export const ListCampaignsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  targetUrl: zod.string(),
  competitorUrls: zod.array(zod.string()),
  seedQueries: zod.array(zod.string()),
  status: zod.enum(["idle", "running", "completed", "failed"]),
  createdAt: zod.string(),
  completedAt: zod.string().nullish(),
});
export const ListCampaignsResponse = zod.array(ListCampaignsResponseItem);

/**
 * @summary Create a new campaign
 */
export const CreateCampaignBody = zod.object({
  name: zod.string(),
  targetUrl: zod.string(),
  competitorUrls: zod.array(zod.string()),
  seedQueries: zod.array(zod.string()),
});

/**
 * @summary Get a campaign by ID
 */
export const GetCampaignParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCampaignResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  targetUrl: zod.string(),
  competitorUrls: zod.array(zod.string()),
  seedQueries: zod.array(zod.string()),
  status: zod.enum(["idle", "running", "completed", "failed"]),
  createdAt: zod.string(),
  completedAt: zod.string().nullish(),
});

/**
 * @summary Delete a campaign
 */
export const DeleteCampaignParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteCampaignResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Run the campaign analysis (SSE stream)
 */
export const RunCampaignParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get stored LLM responses for a campaign
 */
export const GetCampaignResultsParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCampaignResultsResponseItem = zod.object({
  id: zod.number(),
  campaignId: zod.number(),
  llm: zod.string(),
  query: zod.string(),
  responseText: zod.string(),
  createdAt: zod.string(),
});
export const GetCampaignResultsResponse = zod.array(
  GetCampaignResultsResponseItem,
);

/**
 * @summary Get computed AEO metrics report for a campaign
 */
export const GetCampaignReportParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCampaignReportResponse = zod.object({
  campaignId: zod.number(),
  totalQueries: zod.number(),
  totalResponses: zod.number(),
  brandStats: zod.array(
    zod.object({
      brand: zod.string(),
      url: zod.string(),
      mentions: zod.number(),
      totalMentions: zod.number(),
      shareOfVoice: zod.number(),
      avgRankPosition: zod.number().nullish(),
      isTarget: zod.boolean(),
    }),
  ),
  llmVisibility: zod.array(
    zod.object({
      llm: zod.string(),
      targetMentions: zod.number(),
      totalMentions: zod.number(),
      shareOfVoice: zod.number(),
      competitorBreakdown: zod.array(
        zod.object({
          brand: zod.string(),
          mentions: zod.number(),
        }),
      ),
    }),
  ),
});