# lib/api-zod/src/generated/types/aeoReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BrandMentionStat } from "./brandMentionStat";
import type { LlmVisibility } from "./llmVisibility";

export interface AeoReport {
  campaignId: number;
  totalQueries: number;
  totalResponses: number;
  brandStats: BrandMentionStat[];
  llmVisibility: LlmVisibility[];
}