# lib/api-zod/src/generated/types/brandMentionStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BrandMentionStat {
  brand: string;
  url: string;
  mentions: number;
  totalMentions: number;
  shareOfVoice: number;
  avgRankPosition?: number | null;
  isTarget: boolean;
}