# lib/api-zod/src/generated/types/campaign.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignStatus } from "./campaignStatus";

export interface Campaign {
  id: number;
  name: string;
  targetUrl: string;
  competitorUrls: string[];
  seedQueries: string[];
  status: CampaignStatus;
  createdAt: string;
  completedAt?: string | null;
}