# lib/api-zod/src/generated/types/createCampaignBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCampaignBody {
  name: string;
  targetUrl: string;
  competitorUrls: string[];
  seedQueries: string[];
}