# lib/api-zod/src/generated/types/llmResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LlmResponse {
  id: number;
  campaignId: number;
  llm: string;
  query: string;
  responseText: string;
  createdAt: string;
}