# lib/api-zod/src/generated/types/llmVisibility.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AEO Analytics Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { LlmVisibilityCompetitorBreakdownItem } from "./llmVisibilityCompetitorBreakdownItem";

export interface LlmVisibility {
  llm: string;
  targetMentions: number;
  totalMentions: number;
  shareOfVoice: number;
  competitorBreakdown: LlmVisibilityCompetitorBreakdownItem[];
}